Added express scaffold task
Scaffold.js is a very helpful tool created by my former classmate,
Shawn Fraistat.
scaffold.js is a program that functions similar to the scaffold tool in
Ruby on Rails. It allows you to generate a scaffold of a resource from
the command line, creating a new route handler, curl scripts and an
updated server.js.

- The `games` resource will be generated using this tool.