Implement /leaderboard route with totalScore
`totalScore` is a field for Users that will accumulate all of the points
the user has scored through all games.

The leaderboard route is a GET action that will return a list of players
by their score.
This action will populate the query with the totalScore field and sort
the list of players returned descending by their totalScore